+++ ShopWithMe/Services/SyncDebugLogger.swift
@@ -33,6 +33,7 @@ enum SyncDebugLogger {
         case snapshotEmpfangen = "sync_snapshot_empfangen"
         case ordnerZugriffFehlgeschlagen = "sync_ordner_zugriff_fehlgeschlagen"
         case baumelndeReferenzGefunden = "sync_baumelnde_referenz_gefunden"
+        case einkaufslistenStand = "sync_einkaufslisten_stand"
         case debugModeEnabled = "debug_mode_enabled"
         case debugModeDisabled = "debug_mode_disabled"
     }